requirements_jax.txt
# JAX/Flax Dependencies for TPU v4-64

# Core JAX (TPU support)
jax[tpu]>=0.4.20
flax>=0.7.5
optax>=0.1.7
orbax-checkpoint>=0.5.0
chex>=0.1.85

# Utilities
einops>=0.7.0
tqdm
pydantic>=2.0
omegaconf
hydra-core
wandb

# Data processing
numpy
huggingface_hub
packaging

# Development/testing
pytest
